requirements.txt
Flask
Flask-PyMongo 
pytest
pytest-cov
pytest-html
mongomock
python-dotenv
Flask-Session
pytest-mock
pycryptodome